Aesthetic Changes / Added Instagram + Button
Changes to the Square backgrounds to add colours. I have added a Give treats button and the Instagram page into one of the squares. Improved the modal design with colours.

To do - Add content to the modal / finish off.

Add modal to give treats and make it functional.

Maybe remove the content from Square two and add the picture to the background (faded)